requirements.txt
discord.py==2.5.2
openai==0.28
python-dotenv==1.0.0
elevenlabs==2.6.0